fixed int error

+++ src/carsimulator/carui/Grid.java
@@ -122,7 +122,7 @@ public class Grid extends JFrame implements Runnable {
         g.drawImage(this.img, 0, 0, this);
         g.drawImage(this.img2_rotated, this.carModel.getLocation().x, this.carModel.getLocation().y, this);
         g.drawString("Gas: " + Integer.toString(carModel.getCurrentGasAmount()) + " / 100", 800, 50);
-        g.drawString("Speed: " + Integer.toString(carModel.getSpeed()), 800, 100);
+        g.drawString("Speed: " + Double.toString(carModel.getSpeed()), 800, 100);
         g.drawString("Internal Temperature: " + Double.toString(carModel.getInternalTemperature()), 800, 150);
         g.drawString("Engine Temperature: " + Double.toString(carModel.getEngineTemperature()), 800, 200);
